refactor(provider): extract ACP adapter factory
